+++ src/Methods/ModelForwardsCallsExtension.php
@@ -4,14 +4,8 @@ declare(strict_types=1);
 
 namespace NunoMaduro\Larastan\Methods;
 
-use Illuminate\Database\Eloquent\Builder as EloquentBuilder;
 use Illuminate\Database\Eloquent\Model;
-use Illuminate\Database\Query\Builder as QueryBuilder;
-use Illuminate\Support\Str;
-use NunoMaduro\Larastan\Concerns;
-use PHPStan\Reflection\BrokerAwareExtension;
 use PHPStan\Reflection\ClassReflection;
-use PHPStan\Reflection\Dummy\DummyMethodReflection;
 use PHPStan\Reflection\MethodReflection;
 use PHPStan\Reflection\MethodsClassReflectionExtension;
 use PHPStan\Reflection\MissingMethodFromReflectionException;
@@ -19,13 +13,14 @@ use PHPStan\ShouldNotHappenException;
 use PHPStan\Type\Generic\GenericObjectType;
 use PHPStan\Type\ObjectType;
 
-final class ModelForwardsCallsExtension implements MethodsClassReflectionExtension, BrokerAwareExtension
+final class ModelForwardsCallsExtension implements MethodsClassReflectionExtension
 {
-    use Concerns\HasBroker;
+    /** @var BuilderHelper */
+    private $builderHelper;
 
-    private function getBuilderReflection(): ClassReflection
+    public function __construct(BuilderHelper $builderHelper)
     {
-        return $this->broker->getClass(EloquentBuilder::class);
+        $this->builderHelper = $builderHelper;
     }
 
     public function hasMethod(ClassReflection $classReflection, string $methodName): bool
@@ -34,26 +29,16 @@ final class ModelForwardsCallsExtension implements MethodsClassReflectionExtensi
             return false;
         }
 
-        if (in_array($methodName, ['increment', 'decrement', 'paginate', 'simplePaginate'], true)) {
-            return true;
-        }
+        $customBuilderName = $this->builderHelper->determineBuilderType($classReflection->getName());
 
-        // Model scopes
-        if ($classReflection->hasNativeMethod('scope'.ucfirst($methodName))) {
-            return true;
-        }
+        $returnMethodReflection = $this->builderHelper->getMethodReflectionFromBuilder(
+            $classReflection,
+            $methodName,
+            $classReflection->getName(),
+            new GenericObjectType($customBuilderName, [new ObjectType($classReflection->getName())])
+        );
 
-        // Dynamic wheres
-        if (Str::startsWith($methodName, 'where')) {
-            return true;
-        }
-
-        $builderHelper = new BuilderHelper($this->getBroker());
-        $customBuilder = $builderHelper->determineBuilderType($classReflection->getName());
-
-        return $this->getBuilderReflection()->hasNativeMethod($methodName)
-            || $this->broker->getClass(QueryBuilder::class)->hasNativeMethod($methodName)
-            || ($customBuilder && $this->broker->getClass($customBuilder)->hasNativeMethod($methodName));
+        return $returnMethodReflection !== null;
     }
 
     /**
@@ -66,20 +51,19 @@ final class ModelForwardsCallsExtension implements MethodsClassReflectionExtensi
      */
     public function getMethod(ClassReflection $originalModelReflection, string $methodName): MethodReflection
     {
-        $builderHelper = new BuilderHelper($this->getBroker());
-        $customBuilderName = $builderHelper->determineBuilderType($originalModelReflection->getName());
+        $customBuilderName = $this->builderHelper->determineBuilderType($originalModelReflection->getName());
 
-        $returnMethodReflection = $builderHelper->getMethodReflectionFromBuilder(
+        $returnMethodReflection = $this->builderHelper->getMethodReflectionFromBuilder(
             $originalModelReflection,
             $methodName,
             $originalModelReflection->getName(),
             new GenericObjectType($customBuilderName, [new ObjectType($originalModelReflection->getName())])
         );
 
-        if ($returnMethodReflection !== null) {
-            return $returnMethodReflection;
+        if ($returnMethodReflection === null) {
+            throw new ShouldNotHappenException();
         }
 
-        return new DummyMethodReflection($methodName);
+        return $returnMethodReflection;
     }
 }