+++ src/Methods/RelationForwardsCallsExtension.php
@@ -5,18 +5,22 @@ declare(strict_types=1);
 namespace NunoMaduro\Larastan\Methods;
 
 use Illuminate\Database\Eloquent\Relations\Relation;
-use NunoMaduro\Larastan\Concerns\HasBroker;
-use PHPStan\Reflection\BrokerAwareExtension;
 use PHPStan\Reflection\ClassReflection;
-use PHPStan\Reflection\Dummy\DummyMethodReflection;
 use PHPStan\Reflection\MethodReflection;
 use PHPStan\Reflection\MethodsClassReflectionExtension;
+use PHPStan\ShouldNotHappenException;
 use PHPStan\Type\Generic\GenericObjectType;
 use PHPStan\Type\ObjectType;
 
-final class RelationForwardsCallsExtension implements MethodsClassReflectionExtension, BrokerAwareExtension
+final class RelationForwardsCallsExtension implements MethodsClassReflectionExtension
 {
-    use HasBroker;
+    /** @var BuilderHelper */
+    private $builderHelper;
+
+    public function __construct(BuilderHelper $builderHelper)
+    {
+        $this->builderHelper = $builderHelper;
+    }
 
     public function hasMethod(ClassReflection $classReflection, string $methodName): bool
     {
@@ -24,33 +28,45 @@ final class RelationForwardsCallsExtension implements MethodsClassReflectionExte
             return false;
         }
 
-        return $classReflection->getActiveTemplateTypeMap()->getType('TRelatedModel') !== null;
-    }
-
-    public function getMethod(
-        ClassReflection $classReflection,
-        string $methodName
-    ): MethodReflection {
-        $builderHelper = new BuilderHelper($this->getBroker());
-
         /** @var ObjectType|null $relatedModel */
         $relatedModel = $classReflection->getActiveTemplateTypeMap()->getType('TRelatedModel');
 
         if ($relatedModel === null) {
-            return new DummyMethodReflection($methodName);
+            return false;
         }
 
-        $returnMethodReflection = $builderHelper->getMethodReflectionFromBuilder(
+        $returnMethodReflection = $this->builderHelper->getMethodReflectionFromBuilder(
             $classReflection,
             $methodName,
             $relatedModel->getClassName(),
             new GenericObjectType($classReflection->getName(), [$relatedModel])
         );
 
-        if ($returnMethodReflection !== null) {
-            return $returnMethodReflection;
+        return $returnMethodReflection !== null;
+    }
+
+    public function getMethod(
+        ClassReflection $classReflection,
+        string $methodName
+    ): MethodReflection {
+        /** @var ObjectType|null $relatedModel */
+        $relatedModel = $classReflection->getActiveTemplateTypeMap()->getType('TRelatedModel');
+
+        if ($relatedModel === null) {
+            throw new ShouldNotHappenException(sprintf('%s does not have TRelatedModel template type. But it should.', $classReflection->getName()));
         }
 
-        return new DummyMethodReflection($methodName);
+        $returnMethodReflection = $this->builderHelper->getMethodReflectionFromBuilder(
+            $classReflection,
+            $methodName,
+            $relatedModel->getClassName(),
+            new GenericObjectType($classReflection->getName(), [$relatedModel])
+        );
+
+        if ($returnMethodReflection === null) {
+            throw new ShouldNotHappenException(sprintf('%s does not have %s method. But it should.', $classReflection->getName(), $methodName));
+        }
+
+        return $returnMethodReflection;
     }
 }